+++ example/lib/custom_dialog_agent.dart
@@ -1,22 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 
-const customAgentName = 'customDialogAgent';
-const customDialogName = 'customDialog';
+int? _customDialogId;
 
 class CustomDialogAgent extends EasyDialogAgentBase {
   CustomDialogAgent({required super.overlayController});
 
   @override
   Future<void> hide({AgentHideParams? params}) async {
-    super.overlayController.removeCustomDialog(name: customDialogName);
+    super.overlayController.removeCustomDialog(_customDialogId!);
   }
 
   @override
   Future<void> show({required CustomAgentShowParams params}) async {
-    super.overlayController.insertCustomDialog(
-          name: customDialogName,
-          dialog: Align(
+    _customDialogId = super.overlayController.insertCustomDialog(
+          Align(
             alignment: Alignment.bottomCenter,
             child: Container(
               height: 200.0,